refactor: 변경된 도메인 적용
